pyproject.toml
[tool.poetry]
name          = "smol_k8s_lab"
version       = "1.0.0"
description   = "bootstrap simple projects on kubernetes with kind and k3s"
authors       = ["Jesse Hitch <jessebot@linux.com>",
                 "Max Roby <emax@cloudydev.net>"]
readme        = "README.md"
packages      = [{include = "smol_k8s_lab"}]
license       = "AGPL-3.0-or-later"
homepage      = "https://small-hack.github.io/smol-k8s-lab"
repository    = "http://github.com/small-hack/smol-k8s-lab"
documentation = "https://small-hack.github.io/smol-k8s-lab"
keywords      = ["kubernetes", "homelab", "kind", "k3s", "k8s"]
classifiers   = ["Development Status :: 3 - Alpha",
                 "Programming Language :: Python :: 3.11",
                 "Operating System :: MacOS :: MacOS X",
                 "Operating System :: POSIX :: Linux",
                 "Intended Audience :: End Users/Desktop",
                 "Topic :: System :: Installation/Setup",
                 "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"]
include       = ["smol_k8s_lab/config/kind_cluster_config.yaml",
                 "smol_k8s_lab/config/extras/Brewfile",
                 "smol_k8s_lab/config/default_config.yaml"]

[tool.poetry.urls]
"Bug Tracker" = "http://github.com/small-hack/smol-k8s-lab/issues"

[tool.poetry.dependencies]
python        = ">=3.11,<4.0"
rich          = "^13.0"
click         = "^8.1"
bcrypt        = "^4.0"
kubernetes    = "^27.2"
requests      = "^2.28"
pyyaml        = "^6.0"
xdg-base-dirs = "^6.0"
pyjwt = "^2.8.0"
cryptography = "^41.0.3"

[tool.poetry.plugins."smol-k8s-lab.application.plugin"]
"smol-k8s-lab" = "smol_k8s_lab:main"

[tool.poetry.scripts]
smol-k8s-lab = "smol_k8s_lab:main"

[tool.poetry.group.dev.dependencies]
deptry = "^0.12.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
